requirements.txt
fastapi
uvicorn
torch
torchvision
torchaudio
open_clip_torch
faiss-cpu
pillow
numpy
